Update Api và fix lỗi

+++ app/Repositories/ProductImageRepository.php
@@ -4,6 +4,8 @@ namespace App\Repositories;
 
 use App\Models\Product;
 use App\Models\ProductImage;
+use Illuminate\Support\Facades\Storage;
+
 class ProductImageRepository
 {
     public function findProduct($productId)
@@ -19,7 +21,7 @@ class ProductImageRepository
         return ProductImage::create([
             'product_id' => $productId,
             'image_url' => $path,
-            'is_main' => $isMain,
+            'is_main' => $isMain ? 1 : 0,
         ]);
     }
 
@@ -36,7 +38,20 @@ class ProductImageRepository
      */
     public function deleteImage(ProductImage $image)
     {
-        Storage::disk('public')->delete($image->image_url);
+        if (Storage::disk('public')->exists($image->image_url))
+        {
+            Storage::disk('public')->delete($image->image_url);
+        }
+
+        if ($image->is_main) {
+            $nextImage = ProductImage::where('product_id', $image->product_id)
+            ->where('id', '!=', $image->id)
+            ->first();
+
+            if ($nextImage) {
+                $nextImage->update(['is_main' => true]);
+            }
+        }
         return $image->delete();
     }
 
@@ -45,7 +60,7 @@ class ProductImageRepository
      */
     public function getImagesByProduct($productId)
     {
-        return ProductImage::where('product_id', $productId)->get();
+        return ProductImage::where('product_id', $productId)->orderByDesc('is_main')->get();
     }
 
 }
